simplify logic for determining files to crop
